job-assistant/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.8
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
openai==1.3.7
anthropic==0.7.8
selenium==4.15.2
webdriver-manager==4.0.1
schedule==1.2.0
python-telegram-bot==20.7
smtplib-ssl==0.0.1
email-validator==2.1.0
jinja2==3.1.2
aiofiles==23.2.1
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.0.3
celery==5.3.4
redis==5.0.1
linkedin-api==2.0.0
pandas==2.1.3
numpy==1.25.2